backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.9
pydantic==2.9.0
python-dotenv==1.0.1
supabase==2.7.1
python-jose[cryptography]==3.3.0